Update GRIP-E results for GEM-Hydro
Etienne:
Instead of the Manning values calibrated when using MESH-SVS + Watroute (which are fixed all over the domain),
I simply used the default, spatially-varying Manning values of GEM-Hydro, as well as activating the default spatially-varying meandering factor (increasing the length of the river in a grid cell) and default Rlake factor (to adjust Manning values to account for the presence of lakes in the grid cell).
In simple words, I used the default GEM-Hydro routing configuration (except I use the calibrated LZS parameters), which cannot be used with MESH+Watroute yet.